requirements.txt
playwright==1.41.0
anthropic==0.18.1
pyyaml==6.0.1
beautifulsoup4==4.12.3
aiohttp==3.9.3
